Logout on all pages
1) Moved logout logic to an export function so that it can be imported in other pages
2) Added logout api call to end the session on firebase on the HandleLogout function
